refactor(vehicle_info): restructure and modularise things a bit
* move brand icon funcs to module
* add vehicle data factory
* add util funcs for vehicle info and MOT embed generation
